backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
google-auth>=2.27.0

# Redis
redis>=5.0.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Development
python-dotenv>=1.0.0
